sysuser: turn UserList into container component
